api/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
